Date: 20200622
File: portage-latest.tar.xz
Verify result:
gpg: Signature made Mon Jun 22 02:56:10 2020 CEST
gpg:                using RSA key E1D6ABB63BFCFB4BA02FDF1CEC590EEAC9189250
gpg: Good signature from "Gentoo Portage Snapshot Signing Key (Automated Signing Key)" [expired]
gpg: Note: This key has expired!
Primary key fingerprint: DCD0 5B71 EAB9 4199 527F  44AC DB6B 8C1F 96D8 BF6D
     Subkey fingerprint: E1D6 ABB6 3BFC FB4B A02F  DF1C EC59 0EEA C918 9250

gpgsig:
-----BEGIN PGP SIGNATURE-----

iQKTBAABCgB9FiEE4dartjv8+0ugL98c7FkO6skYklAFAl7wAapfFIAAAAAALgAo
aXNzdWVyLWZwckBub3RhdGlvbnMub3BlbnBncC5maWZ0aGhvcnNlbWFuLm5ldEUx
RDZBQkI2M0JGQ0ZCNEJBMDJGREYxQ0VDNTkwRUVBQzkxODkyNTAACgkQ7FkO6skY
klBcuRAApbHNoxiipSGmbij+NjvkCzm1a7QLYibfPy6xhd73qVos/p98FoR1/kCw
/UQE+1pwHCV2TJMQEiRdmkzVOm+sCpe/OWCO4NB6eExNoIUrE8LAMtNxrU7IckqC
IOgh6Z13e2UyZ9bMaQYxynt3VhewinRVFM5WSwxFNZLm1yn8hEUcLRN8W4y6oo7E
s0KNzW2eda7NNqZalQ3GfFPCeAWtwO+qPLlY7yh9BsE9U13qBvMKgRFhFjvK5oBc
cXLI7DncJJOA24lr/65zVy18ftSw5bjfNXJMalVIT+TeUBQ8XZRXgGWSicQNgXiC
W6voDQTqsZpOVHZTG0HsIIE9+TArs2MvMugxlcL9G99n78HjV4ag9bapQif0BapH
5Fg8Brsn0kcG7T5n8KBTBb5zFjmOha9QHEJPuSdPf1OFZE4Tpu+kzETo69ORufA0
F/WXPYT19/Y6YdT7x1tWjWZvY5F1LFMr8Da9ysXGBQDw440vEMXvDLQUfm8ad4Sl
2uDBOiYwc6mBQqbBC5JoWifZ0jfpkZHVsDrg/3E8h/ErQSeq77XZfpJjxPzgOT1Z
DH2x+Ijpy78xPWkxAEdRzaL1mUSe0UOSrWSH55TMx/5i1zuf6bbN/O7pk6l9n4gg
gSepB8rh8k/F1gTFCAo0uTj8aTkwCnhaRYu+BcRxJGG7TAb69AI=
=zst+
-----END PGP SIGNATURE-----

+++ app-admin/diamond/diamond-4.0.515.20200620.ebuild
@@ -11,7 +11,7 @@ if [[ ${PV} = 9999* ]]; then
 else
 	GHASH=8d8a2e49d80d44968a34d43e36c1d864695a29c1  # from python3 branch
 	SRC_URI="https://github.com/grobian/Diamond/archive/${GHASH}.tar.gz -> python-diamond-${PV}.tar.gz"
-	KEYWORDS="~amd64 ~x86 ~amd64-linux ~ppc-macos ~x64-macos ~x86-macos ~sparc-solaris ~sparc64-solaris ~x64-solaris ~x86-solaris"
+	KEYWORDS="~amd64 ~arm ~x86 ~amd64-linux ~ppc-macos ~x64-macos ~x86-macos ~sparc-solaris ~sparc64-solaris ~x64-solaris ~x86-solaris"
 	S=${WORKDIR}/Diamond-${GHASH}
 fi
 
@@ -64,7 +64,11 @@ python_install() {
 	export VIRTUAL_ENV=1
 	distutils-r1_python_install
 	python_optimize
-	mv "${ED}"/usr/etc "${ED}"/ || die
+	# since python3.8 installation goes straight into /etc
+	[[ -d ${ED}/etc ]] && [[ -d ${ED}/usr/etc ]] && rm -Rf "${ED}"/usr/etc
+	if [[ -d ${ED}/usr/etc ]] ; then
+		mv "${ED}"/usr/etc "${ED}"/ || die
+	fi
 	rm "${ED}"/etc/diamond/*.windows  # won't need these
 	sed -i \
 		-e '/pid_file =/s:/var/run:/run:' \
